Hardcode info record location for dry run
At dry-run time, the NanoKernel is not asked for the location of HWInfo.
The pointer block at 5FFFEFC0 is used instead. The location of
ConfigInfo was already hardcoded.

This neatens things up considerably, but will cause the dry-run Wedge to
crash on a machine with a Wedged address space. If the
NKLocateInfoRecord code can be cleaned up a bit, then it might be
restored.